Removed unnecessary items.
"Powered by" and "Check out the GitHub repository" are not necessary and give an unprofessional look to the website. The website is what the users should look at, so there is no reason to provide a link to the code hosting the website on every page, or any page at all.

As for the "Powered by" part, it is not powered by a sponsor, so it's not necessary either.